Move models to pretixapi
