+++ src/pretix/api/__init__.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class PretixApiConfig(AppConfig):
+    name = 'pretix.api'
+    label = 'pretixapi'
+
+
+default_app_config = 'pretix.api.PretixApiConfig'
