+++ src/pretix/base/models/__init__.py
@@ -1,8 +1,5 @@
 from ..settings import GlobalSettingsObject_SettingsStore
-from .auth import (
-    OAuthAccessToken, OAuthApplication, OAuthGrant, OAuthRefreshToken,
-    U2FDevice, User,
-)
+from .auth import U2FDevice, User
 from .base import CachedFile, LoggedModel, cachedfile_name
 from .checkin import Checkin, CheckinList
 from .event import (
